Made payload type int in jitter buffers.
Added array of jitter buffers to MprDejitter. They are initialized when we select codecs in MpRtpInputAudioConnection - before 1st rtp frame is received.

+++ trunk/sipXmediaLib/include/mp/MpJitterBufferBase.h
@@ -58,7 +58,7 @@ public:
     *        80, 160, 240...
     */
    MpJitterBufferBase(const UtlString& name,
-                      uint8_t payloadType,
+                      int payloadType,
                       unsigned int frameSize);
 
    virtual ~MpJitterBufferBase();
@@ -100,7 +100,7 @@ public:
 ///@name Accessors
 //@{
 
-   uint8_t getPayloadType();
+   int getPayloadType();
 
 //@}
 
@@ -113,7 +113,7 @@ public:
 /* //////////////////////////// PROTECTED ///////////////////////////////// */
 protected:
    UtlString m_name;
-   uint8_t m_payloadType;
+   int m_payloadType;
    unsigned int m_frameSize;
 
 /* //////////////////////////// PRIVATE /////////////////////////////////// */
